+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fix
 * Handle case when input sensorchan strings are string subclasses.
+* Fix issue where lazy warps did not respect explicitly given dsize arguments
 
 
 ## Version 0.4.5 - Released 2025-04-09